# backend/docs/api-spec.yaml
openapi: 3.0.0
info:
  title: Felparkering API
  description: Rest API för att hantera felparkerade fordon
  version: '1.0'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    variables: {}
paths:
...
  /reports:
    post:
      tags:
        - Reports
      summary: Create new report
      description: |-
        ### POST /reports

        Skapar en felparkeringsrapport.

        #### Behörighet:

        - `ADMIN`: Ja
            
        - `ATTENDANT`: Nej (403)
            
        - `CUSTOMER`: Ja
      operationId: Createnewreport
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  example: 101
                street:
                  type: string
                  example: Södra vägen
                houseNumber:
                  type: string
                  example: "1"
                city:
                  type: string
                  example: Stockholm
                licensePlate:
                  type: string
                  example: ABC123
                category:
                  type: string
                  example: NO_PARKING_AREA
      responses:
        '201':
          description: Created
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
              example:
                data:
                  id: 12
                  address:
                    id: 101
                    longitude: 18.06
                    latitude: 59.33
                    street: Södra vägen
                    houseNumbers:
                      - "1"
                    city: Stockholm
                    distanceFromCity: 120
                  licensePlate: ABC123
                  category: NO_PARKING_AREA
                  attendantGroup:
                    id: 1
                    name: Stockholm Södra
                  assignedToId: null
                  createdOn: '2025-06-17T14:22:00Z'
                  updatedOn: '2025-06-17T14:22:00Z'
                  status: NEW
                message: Report created successfully
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
              example:
                errors:
                  - field: licensePlate
                    message: License plate is required
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: ACCESS_DENIED
                  message: You do not have permission to get all reports
      deprecated: false
      security:
        - bearer: []
    get:
      tags:
        - Reports
      summary: Get all reports
      description: |-
        ### GET /reports

        Hämtar en lista med alla felparkeringsrapporter.

        #### Behörighet:

        Roll:

        - `ADMIN`: Ser alla rapporter
            
        - `ATTENDANT`: Ser endast rapporter för sin egen vaktgrupp
            
        - `CUSTOMER`: Ser endast rapporter som de har skapat
      operationId: Getallreports
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
        - name: sortBy
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: createdOn
        - name: sortDir
          in: query
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - NEW
              - ASSIGNED
              - RESOLVED
              - CANCELLED
        - name: assignedToId
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedReportsResponse'
              example:
                data:
                  items:
                    - id: 12
                      address:
                        id: 101
                        street: Södra vägen
                        houseNumbers:
                          - "1"
                        city: Stockholm
                        latitude: 59.33
                        longitude: 18.06
                        distanceFromCity: 120
                      licensePlate: ABC123
                      category: NO_PARKING_AREA
                      attendantGroup:
                        id: 1
                        name: Stockholm Södra
                      assignedToId: 7
                      createdOn: '2025-06-17T14:22:00Z'
                      updatedOn: '2025-06-17T14:22:00Z'
                      status: NEW
                  page: 0
                  size: 10
                  totalElements: 1
                  totalPages: 1
                message: Reports fetched successfully
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: ACCESS_DENIED
                  message: You do not have permission to get all reports
      deprecated: false
      security:
        - bearer: []
  /reports/{id}:
    get:
      tags:
        - Reports
      summary: Get a report
      description: |-
        ### GET /reports/{id}

        Hämtar detaljerad information om en specifik felparkeringsrapport.

        #### Behörighet:

        - `ADMIN`: Har tillgång till alla rapporter
            
        - `ATTENDANT`: Har tillgång om rapporten tillhör samma vaktgrupp
            
        - `CUSTOMER`: Har tillgång om rapporten är skapad av användaren
            

        #### Parameter

        - `id` _(Long)_ – rapportens ID
      operationId: Getareport
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: string
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
              example:
                data:
                  id: 12
                  address:
                    id: 101
                    street: Södra vägen
                    houseNumbers:
                      - "1"
                    city: Stad
                    latitude: 59.3293
                    longitude: 18.0686
                    distanceFromCity: 120
                  licensePlate: ABC123
                  category: NO_PARKING_AREA
                  attendantGroup:
                    id: 1
                    name: Stockholm Södra
                  assignedToId: 7
                  createdOn: '2025-06-17T14:22:00Z'
                  updatedOn: '2025-06-17T14:22:00Z'
                  status: NEW
                message: Report fetched successfully
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: ACCESS_DENIED
                  message: You do not have permission to get all reports
        '404':
          description: Not Found
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: Report not found
      deprecated: false
      security:
        - bearer: []
    put:
      tags:
        - Reports
      summary: Update report status
      description: |-
        ### PUT/reports/{id}

        Uppdaterar informationen för en specifik felanmälan, t.ex. ändrar dess status.

        #### Behörighet:

        - `ADMIN`: kan uppdatera status på alla rapporter
            
        - `ATTENDANT`: kan ändra status om rapporten tillhör deras vaktgrupp
            
        - `CUSTOMER`: kan ändra status på sina egna ärenden om de är aktiva till `CANCELLED`
            

        #### Parameter

        - `id` _(Long)_ – rapportens ID
      operationId: Updatereportstatus
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum:
                    - NEW
                    - ASSIGNED
                    - RESOLVED
                    - CANCELLED
            example:
              status: RESOLVED
      responses:
        '200':
          description: Updated
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
              example:
                data:
                  id: 12
                  address:
                    id: 101
                    longitude: 18.06
                    latitude: 59.33
                    street: Södra vägen
                    houseNumbers:
                      - "1"
                    city: Stockholm
                    distanceFromCity: 120
                  licensePlate: ABC123
                  category: NO_PARKING_AREA
                  attendantGroup:
                    id: 1
                    name: Stockholm Södra
                  assignedToId: null
                  createdOn: '2025-06-17T14:22:00Z'
                  updatedOn: '2025-06-17T14:22:00Z'
                  status: RESOLVED
                message: Report updated successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorResponse'
                  - $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                invalidStatus:
                  summary: Invalid status value
                  value:
                    error:
                      code: BAD_REQUEST
                      message: Invalid request parameter or value
                missingStatus:
                  summary: Missing status
                  value:
                    errors:
                      - field: status
                        message: must not be null
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: ACCESS_DENIED
                  message: You do not have permission to get all reports
        '404':
          description: Not Found
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: Report not found
        '409':
          description: Conflict
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: CONFLICT
                  message: Report is already assigned to a different user
      deprecated: false
      security:
        - bearer: []
...
components:
  schemas:
...
    ValidationErrorResponse:
      title: ValidationErrorResponse
      required:
        - errors
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
              message:
                type: string
      example:
        errors:
          - field: status
            message: must not be null
    ErrorResponse:
      title: ErrorResponse
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      example:
        error:
          code: BAD_REQUEST
          message: Bad request
    ErrorBody:
      title: ErrorBody
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
      example:
        code: BAD_REQUEST
        message: Bad request
...
    PagedReportResponse:
      title: PagedReportResponse
      required:
        - items
        - page
        - size
        - totalElements
        - totalPages
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReportDetailDto'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
      example:
        items:
          - id: 12
            address:
              id: 101
              street: Södra vägen
              houseNumbers:
                - "1"
              city: Stockholm
              latitude: 59.33
              longitude: 18.06
              distanceFromCity: 120
            licensePlate: ABC123
            category: NO_PARKING_AREA
            attendantGroup:
              id: 1
              name: Stockholm Södra
            assignedToId: 7
            createdOn: '2025-06-17T14:22:00Z'
            updatedOn: '2025-06-17T14:22:00Z'
            status: NEW
        page: 0
        size: 10
        totalElements: 1
        totalPages: 1
    PagedReportsResponse:
      title: PagedReportsResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PagedReportResponse'
        message:
          type: string
      example:
        data:
          items: []
          page: 0
          size: 10
          totalElements: 0
          totalPages: 0
        message: Reports fetched successfully
    ReportResponse:
      title: ReportResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReportDetailDto'
        message:
          type: string
      example:
        data:
          id: 12
          address:
            id: 101
            longitude: 18.06
            latitude: 59.33
            street: Södra vägen
            houseNumbers:
              - "1"
            city: Stockholm
            distanceFromCity: 120
          licensePlate: ABC123
          category: NO_PARKING_AREA
          attendantGroup:
            id: 1
            name: Stockholm Södra
          assignedToId: null
          createdOn: '2025-06-17T14:22:00Z'
          updatedOn: '2025-06-17T14:22:00Z'
          status: NEW
        message: Report fetched successfully
    AttendantGroupDto:
      title: AttendantGroupDto
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
      example:
        id: 1
        name: Stockholm Södra
    AddressDto:
      title: AddressDto
      required:
        - id
        - longitude
        - latitude
        - street
        - houseNumbers
        - city
        - distanceFromCity
      type: object
      properties:
        id:
          type: integer
          format: int64
        longitude:
          type: number
          format: double
        latitude:
          type: number
          format: double
        street:
          type: string
        houseNumbers:
          type: array
          items:
            type: string
        city:
          type: string
        distanceFromCity:
          type: number
          format: double
      example:
        id: 101
        longitude: 18.06
        latitude: 59.33
        street: Södra vägen
        houseNumbers:
          - "1"
        city: Stockholm
        distanceFromCity: 120
...
    ReportDetailDto:
      title: ReportDetailDto
      required:
        - id
        - address
        - licensePlate
        - category
        - attendantGroup
        - assignedToId
        - createdOn
        - updatedOn
        - status
      type: object
      properties:
        id:
          type: integer
          format: int64
        address:
          $ref: '#/components/schemas/AddressDto'
        licensePlate:
          type: string
        category:
          type: string
        attendantGroup:
          $ref: '#/components/schemas/AttendantGroupDto'
        assignedToId:
          type: integer
          format: int64
          nullable: true
          description: ID of the attendant assigned to the report. Null if unassigned.
        createdOn:
          type: string
          format: date-time
        updatedOn:
          type: string
          format: date-time
        status:
          type: string
      example:
        id: 12
        address:
          id: 101
          street: Södra vägen
          houseNumbers:
            - "1"
          city: Stockholm
          latitude: 59.33
          longitude: 18.06
          distanceFromCity: 120
        licensePlate: ABC123
        category: BLOCKERAR_UTFART
        attendantGroup:
          id: 1
          name: Stockholm Södra
        assignedToId: 7
        createdOn: '2025-06-17T14:22:00Z'
        updatedOn: '2025-06-17T14:22:00Z'
        status: NEW
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...